# Fix range locking in ZIL commit codepath