@loli10K 2017-08-11
UPDATE: it seems this issue has been introduced (or it's more likely to be triggered) with _OpenZFS 7578 - Fix/improve some aspects of ZIL writing_ (1b7c1e5).

When we have a ZVOL with `logbias=throughput` we force `itx->itx_wr_state` to `WR_INDIRECT` and the `itx->itx_lr` offset and length to the offset and length of the BIO from `zvol_write()->zvol_log_write()`: these offset and length are later used to take a range lock in `zillog->zl_get_data`/`zvol_get_data()`. The length is actually set to `MIN(blocksize - P2PHASE(offset, blocksize), size)` but i don't think this really changes anything in this context.

Now suppose we have a ZVOL with blocksize=8K and push 4K writes to offset 0: we will only be range-locking 0-4096. This means the ASSERTion we make in `dbuf_unoverride()` is no longer valid because now `dmu_sync()` is called from zilog's get_data functions holding a **partial** lock on the dbuf.

Even though it was initially thought this was somehow related to some combination of virtualization settings and storage drivers this is actually reproducible on a file-based vdev pool with `fio` (or even `dd`), we just have to push writes on the same ZVOL device is a small range (0-8K) fast enough to trigger this race condition.

```
root@linux:~# POOLNAME="poolname"
root@linux:~# TMPDIR='/var/tmp'
root@linux:~# #
root@linux:~# mountpoint -q $TMPDIR || mount -t tmpfs tmpfs $TMPDIR
root@linux:~# zpool destroy $POOLNAME
cannot open 'poolname': no such pool
root@linux:~# rm -f $TMPDIR/zpool.dat
root@linux:~# fallocate -l 128m $TMPDIR/zpool.dat
root@linux:~# zpool create $POOLNAME $TMPDIR/zpool.dat
root@linux:~# zfs create -V 64M -s -o sync=always -o logbias=throughput -b 128K $POOLNAME/zvol
root@linux:~# #
root@linux:~# fio --filename=/dev/zvol/$POOLNAME/zvol --sync=1 --rw=write --bs=8k --size=8k --io_limit=1m --numjobs=2 --name=issue-6238
issue-6238: (g=0): rw=write, bs=8K-8K/8K-8K/8K-8K, ioengine=sync, iodepth=1
...
fio-2.1.11
Starting 2 processes
[   89.823158] VERIFY(dr->dt.dl.dr_override_state != DR_IN_DMU_SYNC) failed
[   89.824208] PANIC at dbuf.c:1308:dbuf_unoverride()

Message from syslogd@linux at Aug 11 22:02:55 ...
 kernel:[   89.823158] VERIFY(dr->dt.dl.dr_override_state != DR_IN_DMU_SYNC) failed

Message from syslogd@linux at Aug 11 22:02:55 ...
 kernel:[   89.824208] PANIC at dbuf.c:1308:dbuf_unoverride()
^ZJobs: 2 (f=2): [W(2)] [7.1% done] [0KB/0KB/0KB /s] [0/0/0 iops] [eta 00m:39s]
[1]+  Stopped                 fio --filename=/dev/zvol/$POOLNAME/zvol --sync=1 --rw=write --bs=8k --size=8k --io_limit=1m --numjobs=2 --name=issue-6238
root@linux:~# dmesg 
[   89.823158] VERIFY(dr->dt.dl.dr_override_state != DR_IN_DMU_SYNC) failed
[   89.824208] PANIC at dbuf.c:1308:dbuf_unoverride()
[   89.824862] Showing stack for process 1930
[   89.824865] CPU: 5 PID: 1930 Comm: zvol Tainted: G           OE   4.6.4 #1
[   89.824866]  0000000000000286 0000000081f2fdd7 ffffffff81315815 ffffffffc07236e0
[   89.824869]  ffff88002a7cbc60 ffffffffc04749e9 ffff88002b6e83c0 0000000000000028
[   89.824871]  ffff88002a7cbc70 ffff88002a7cbc10 6428594649524556 6c642e74643e2d72
[   89.824874] Call Trace:
[   89.824879]  [<ffffffff81315815>] ? dump_stack+0x5c/0x77
[   89.824888]  [<ffffffffc04749e9>] ? spl_panic+0xc9/0x110 [spl]
[   89.824892]  [<ffffffff815ca39e>] ? mutex_lock+0xe/0x30
[   89.824894]  [<ffffffff815ca39e>] ? mutex_lock+0xe/0x30
[   89.824948]  [<ffffffffc06857ce>] ? zio_ready+0x1de/0x610 [zfs]
[   89.824993]  [<ffffffffc068021e>] ? zio_wait+0x20e/0x410 [zfs]
[   89.825022]  [<ffffffffc0582a97>] ? dbuf_unoverride+0x157/0x190 [zfs]
[   89.825050]  [<ffffffffc0582c95>] ? dbuf_redirty+0x35/0xc0 [zfs]
[   89.825079]  [<ffffffffc058d266>] ? dmu_buf_will_dirty+0x96/0x1d0 [zfs]
[   89.825109]  [<ffffffffc0596564>] ? dmu_write_uio_dnode+0x74/0x1d0 [zfs]
[   89.825140]  [<ffffffffc0598778>] ? dmu_write_uio_dbuf+0x48/0x60 [zfs]
[   89.825184]  [<ffffffffc0691a70>] ? zvol_write+0x190/0x600 [zfs]
[   89.825190]  [<ffffffffc04725f3>] ? taskq_thread+0x293/0x580 [spl]
[   89.825193]  [<ffffffff810a5730>] ? wake_up_q+0x60/0x60
[   89.825198]  [<ffffffffc0472360>] ? taskq_thread_spawn+0x50/0x50 [spl]
[   89.825200]  [<ffffffff8109ac4f>] ? kthread+0xdf/0x100
[   89.825203]  [<ffffffff815cc3f2>] ? ret_from_fork+0x22/0x40
[   89.825205]  [<ffffffff8109ab70>] ? kthread_park+0x50/0x50
root@linux:~# 
```

Without `--enable-debug`:

```
root@linux:~# POOLNAME="poolname"
root@linux:~# TMPDIR='/var/tmp'
root@linux:~# #
root@linux:~# mountpoint -q $TMPDIR || mount -t tmpfs tmpfs $TMPDIR
root@linux:~# zpool destroy $POOLNAME
root@linux:~# rm -f $TMPDIR/zpool.dat
root@linux:~# fallocate -l 128m $TMPDIR/zpool.dat
root@linux:~# zpool create $POOLNAME $TMPDIR/zpool.dat
root@linux:~# zfs create -V 64M -s -o sync=always -o logbias=throughput -b 128K $POOLNAME/zvol
root@linux:~# #
root@linux:~# fio --filename=/dev/zvol/$POOLNAME/zvol --sync=1 --rw=write --bs=8k --size=8k --io_limit=1m --numjobs=2 --name=issue-6238
issue-6238: (g=0): rw=write, bs=8K-8K/8K-8K/8K-8K, ioengine=sync, iodepth=1
...
fio-2.1.11
Starting 2 processes
[  232.954702] PANIC: zfs: allocating allocated segment(offset=679424 size=131072)
[  232.954702] 

Message from syslogd@linux at Aug 11 22:33:40 ...
 kernel:[  232.954702] PANIC: zfs: allocating allocated segment(offset=679424 size=131072)

Message from syslogd@linux at Aug 11 22:33:40 ...
 kernel:[  232.954702] 
^Zbs: 2 (f=2)
[1]+  Stopped                 fio --filename=/dev/zvol/$POOLNAME/zvol --sync=1 --rw=write --bs=8k --size=8k --io_limit=1m --numjobs=2 --name=issue-6238
root@linux:~# dmesg
[  232.954702] PANIC: zfs: allocating allocated segment(offset=679424 size=131072)
[  232.955889] Showing stack for process 2013
[  232.955892] CPU: 7 PID: 2013 Comm: txg_sync Tainted: P           OE   4.6.4 #1
[  232.955894]  0000000000000286 0000000067d8f2c9 ffffffff81315815 0000000000000003
[  232.955896]  ffff8800ba27fb90 ffffffffc0450a94 ffffffff810aca93 6c6c61203a73667a
[  232.955897]  20676e697461636f 657461636f6c6c61 6e656d6765732064 74657366666f2874
[  232.955899] Call Trace:
[  232.955903]  [<ffffffff81315815>] ? dump_stack+0x5c/0x77
[  232.955911]  [<ffffffffc0450a94>] ? vcmn_err+0x64/0xf0 [spl]
[  232.955913]  [<ffffffff810aca93>] ? select_idle_sibling+0x23/0x100
[  232.955915]  [<ffffffff810b0833>] ? put_prev_entity+0x33/0x7c0
[  232.955916]  [<ffffffff810b2357>] ? dequeue_task_fair+0x57/0x8e0
[  232.955918]  [<ffffffff810b6a0c>] ? pick_next_task_fair+0x10c/0x4a0
[  232.955920]  [<ffffffff8102b76f>] ? __switch_to+0x26f/0x530
[  232.955954]  [<ffffffffc06ae0a9>] ? zfs_panic_recover+0x69/0x80 [zfs]
[  232.955957]  [<ffffffffc0183141>] ? avl_find+0x51/0x90 [zavl]
[  232.955981]  [<ffffffffc06978af>] ? range_tree_add+0x2af/0x2c0 [zfs]
[  232.956008]  [<ffffffffc06ffaeb>] ? zio_add_child+0xfb/0x110 [zfs]
[  232.956032]  [<ffffffffc06938bc>] ? metaslab_free_dva+0x11c/0x380 [zfs]
[  232.956055]  [<ffffffffc0696178>] ? metaslab_free+0x88/0xd0 [zfs]
[  232.956079]  [<ffffffffc069d840>] ? spa_avz_build+0x130/0x130 [zfs]
[  232.956106]  [<ffffffffc06fd508>] ? zio_dva_free+0x18/0x20 [zfs]
[  232.956133]  [<ffffffffc07017c7>] ? zio_nowait+0xa7/0x140 [zfs]
[  232.956157]  [<ffffffffc069d87b>] ? spa_free_sync_cb+0x3b/0x50 [zfs]
[  232.956174]  [<ffffffffc0653430>] ? bplist_iterate+0x90/0xe0 [zfs]
[  232.956198]  [<ffffffffc06a15d6>] ? spa_sync+0x416/0xd10 [zfs]
[  232.956200]  [<ffffffff810bbfa4>] ? __wake_up+0x34/0x50
[  232.956224]  [<ffffffffc06b1cd1>] ? txg_sync_thread+0x2e1/0x4a0 [zfs]
[  232.956258]  [<ffffffffc06b19f0>] ? txg_sync_stop+0xd0/0xd0 [zfs]
[  232.956264]  [<ffffffffc044c4ae>] ? thread_generic_wrapper+0x7e/0xc0 [spl]
[  232.956269]  [<ffffffffc044c430>] ? __thread_exit+0x20/0x20 [spl]
[  232.956271]  [<ffffffff8109ac4f>] ? kthread+0xdf/0x100
[  232.956273]  [<ffffffff815cc3f2>] ? ret_from_fork+0x22/0x40
[  232.956274]  [<ffffffff8109ab70>] ? kthread_park+0x50/0x50

```

We could fix this by taking a larger range lock in `zvol_get_data()` to protect the whole dbuf before we call `dmu_sync()`. Comments/suggestions are welcome since this is my first time working on ZIL code.